\name{plot.nelson_rule}
\alias{plot.nelson_rule}
\title{plot.nelson_rule}
\usage{
  plot.nelson_rule(x, ...)
}
\arguments{
  \item{x}{object of class /code{nelson_rule}}

  \item{...}{arguments to be passed to methods, such as
  graphical parameters.}
}
\value{
  An object of type
}
\description{
  Creates a control chart with violation points highlighted
  in red.
}

